use base apply_async when task_always_eager=True
Task_always_eager=True is used pretty extensively in tests, and
has some custom logic that celery-singleton is not emulating properly.

As it does not makes much sense to introduce any lock logic when we are
inlining the task code, its more straightforward to just fallback to
the default method whenever we detect task_always_eager is enabled.